Added `setup.load_type_from_package` to reload a single type. Moved `workflow.load_workflow_from_package` to `setup.load_workflow_from_package`. See #PM-3935
